- Merge docket and opinion clone command into one
